+++ src/main/java/com/konrad/garagev3/model/dao/Service.java
@@ -1,7 +1,6 @@
 package com.konrad.garagev3.model.dao;
 
 import javax.persistence.*;
-import java.io.Serializable;
 import java.util.List;
 
 @Entity
@@ -9,7 +8,7 @@ public class Service {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     //@Column(name = "service_id")
-    private int id;
+    private long id;
     @ManyToMany(mappedBy = "services")
     private List<BodyPriceList> bodyPriceLists;
     @ManyToOne(fetch = FetchType.LAZY)
