+++ src/main/java/com/konrad/garagev3/repository/UserRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface UserRepository extends CrudRepository<User, Long> {
     User findByEmail(String email);
 
-    User findUserById(int id);
+    User findUserById(long id);
 
     // TODO: 05.07.2019 Should this method retun eny value
     void deleteUserByEmail(String email);
